chore(logger): rename core 'Logger' to lower case
