+++ packages/core/src/Cache/StorageCacheCommon.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { ConsoleLogger } from '../Logger';
+import { ConsoleLogger } from '../logger';
 import { KeyValueStorageInterface } from '../types';
 import { currentSizeKey, defaultConfig } from './constants';
 import { CacheConfig, CacheItem, CacheItemOptions } from './types';
